Changed from Byte to Short. Should'nt really need to (given my values are only ever between 0-127) but to avoid conflict whatevs -_-

+++ softwaredev/src/simori/MIDIPlayer.java
@@ -34,7 +34,7 @@ public interface MIDIPlayer {
 	 * 
 	 * Method will play all notes simulatenously(or as close to simulatneous as possible)
 	 */
-	public void play(ArrayList<ArrayList<Byte>> array) throws InvalidMidiDataException, InterruptedException;
+	public void play(ArrayList<ArrayList<Short>> array) throws InvalidMidiDataException, InterruptedException;
 
 	
 }
